+++ sibila/schema_format_openai.py
@@ -399,3 +399,195 @@ class FireworksModel(SchemaFormatOpenAIModel):
 
 
 
+
+
+
+class GroqModel(SchemaFormatOpenAIModel):
+    """Access a Groq model with the OpenAI API.
+    Supports constrained JSON output, via the response_format JSON Schema mechanism.
+
+    Ref:
+        https://console.groq.com/docs/tool-use
+
+        https://github.com/groq/groq-api-cookbook/blob/main/parallel-tool-use/parallel-tool-use.ipynb
+    """
+
+    PROVIDER_NAME:str = "groq"
+    """Provider prefix that this class handles."""
+
+    DEFAULT_BASE_URL: str = "https://api.groq.com/openai/v1"
+    """Default API access URL"""
+
+    _token_estimation_factor: float
+    """Multiplication factor to estimate token usage: multiplies text length to obtain token length."""
+
+    DEFAULT_TOKEN_ESTIMATION_FACTOR: float = 0.4
+    """Default factor for token_estimation_factor."""
+
+
+    def __init__(self,
+                 name: str,
+                 *,
+                 
+                 # common base model args
+                 genconf: Optional[GenConf] = None,
+                 schemaconf: Optional[JSchemaConf] = None,
+                 ctx_len: Optional[int] = None,
+                 max_tokens_limit: Optional[int] = None,
+                 tokenizer: Optional[Tokenizer] = None,
+                
+                 # most important OpenAI API specific args
+                 api_key: Optional[str] = None,
+                 base_url: Optional[str] = None,
+                 token_estimation_factor: Optional[float] = None,
+                 
+                 # other OpenAI API specific args
+                 other_init_kwargs: dict = {},
+                 ):
+        """Create a Groq remote model.
+
+        Args:
+            name: Model name to resolve into an existing model.
+            genconf: Model generation configuration. Defaults to None.
+            schemaconf: Default configuration for JSON schema validation, used if generation call doesn't supply one. Defaults to None.
+            ctx_len: Maximum context length to be used (shared for input and output). None for model's default.
+            max_tokens_limit: Maximum output tokens limit. None for model's default.
+            tokenizer: An external initialized tokenizer to use instead of the created from the GGUF file. Defaults to None.
+            api_key: API key. Defaults to None, which will use env variable GROQ_API_KEY.
+            base_url: Base location for API access. Defaults to None, which will use env variable GROK_BASE_URL or a default.
+            token_estimation_factor: Used when no tokenizer is available. Multiplication factor to estimate token usage: multiplies total text length to obtain token length.
+            other_init_kwargs: Extra args for OpenAI.OpenAI() initialization. Defaults to {}.
+
+        Raises:
+            ImportError: If OpenAI API is not installed.
+            NameError: If model name was not found or there's an API or authentication problem.
+        """
+
+        if api_key is None:
+            api_key = os.environ.get("GROQ_API_KEY")
+        if base_url is None:
+            base_url = os.environ.get("GROQ_BASE_URL", self.DEFAULT_BASE_URL)
+
+        super().__init__(name,
+                         # common base model args
+                         genconf=genconf,
+                         schemaconf=schemaconf,
+                         ctx_len=ctx_len,
+                         max_tokens_limit=max_tokens_limit,
+                         tokenizer=tokenizer,
+                            
+                         # most important OpenAI API specific args
+                         api_key=api_key,
+                         base_url=base_url,
+                         token_estimation_factor=token_estimation_factor,
+                            
+                         # other OpenAI API specific args
+                         other_init_kwargs=other_init_kwargs)
+
+
+
+    def _gen_pre(self, 
+                 thread: Thread,
+                 genconf: Union[GenConf, None]
+                 ) -> tuple:
+
+        if genconf is None:
+            genconf = self.genconf
+
+        thread = self._prepare_gen_thread(thread, genconf)
+
+        token_len = self.token_len(thread, genconf)
+        resolved_max_tokens = self.resolve_genconf_max_tokens(token_len, genconf)
+
+
+        json_kwargs: dict = {}
+        if genconf.format == "json":
+            
+            if genconf.json_schema is not None:
+                json_kwargs["tool_choice"] = {
+                    "type": "function",
+                    "function": {"name": self.output_fn_name},
+                }
+
+                if isinstance(genconf.json_schema, str):
+                    params = json.loads(genconf.json_schema)
+                else:
+                    params = genconf.json_schema
+                
+                json_kwargs["tools"] = [
+                    {
+                        "type": "function",
+                        "function": {
+                            "name": self.output_fn_name,
+                            "parameters": params
+                        }
+                    }
+                ]
+
+            else: # free json output
+                json_kwargs["response_format"] = {"type": "json_object"}
+
+
+        # seed config is disabled, remote models and some hardware accelerated local models don't support it.
+        # seed = genconf.seed
+        # if seed == -1:
+        #    seed = int(time())
+        #    logger.debug(f"SchemaFormatOpenAIModel random seed={seed}")
+
+        
+        msgs = thread.as_chatml()
+
+        kwargs = {"model": self._model_name,
+                  "messages": msgs, # type: ignore[arg-type]
+                  "max_tokens": resolved_max_tokens,
+                  "stop": genconf.stop,
+                  "temperature": genconf.temperature,
+                  "top_p": genconf.top_p,
+                  # "seed": seed,
+                  "n": 1,
+                  **json_kwargs}
+
+        # inject model-specific args, if any
+        kwargs.update(genconf.resolve_special(self.PROVIDER_NAME))
+
+        logger.debug(f"{type(self).__name__} gen args: {kwargs}")
+
+        return (kwargs, genconf)
+
+
+    def _gen_post(self, 
+                  response: Any,
+                  pre_kwargs: dict,
+                  genconf: GenConf
+                  ) -> GenOut:
+            
+        logger.debug(f"SchemaFormatOpenAIModel response: {response}")
+
+        choice = response.choices[0]
+        finish = choice.finish_reason
+        # OpenAI-compatible provider endpoints can give non-standard finish_reason values. Map as needed:
+        if finish in ["eos", "tool_calls"]: finish = "stop"
+        message = choice.message
+
+        if "tool_choice" in pre_kwargs:
+            
+            # json schema generation via the tools API:
+            if message.tool_calls is not None:
+                if len(message.tool_calls) != 1:
+                    logger.warn(f"SchemaFormatOpenAIModel: expecting single message.tool_calls, but received {len(message.tool_calls)} - using first.")
+
+                fn = message.tool_calls[0].function
+                if fn.name != self.output_fn_name:
+                    logger.warn(f"SchemaFormatOpenAIModel: expecting '{self.output_fn_name}' function name, received ({fn.name})")
+
+                text = fn.arguments
+
+            else: # use content instead
+                logger.warn("SchemaFormatOpenAIModel: expecting message.tool_calls, but none received - using text content")
+                text = message.content # type: ignore[assignment]
+        
+        else:
+            # text or simple json format
+            text = message.content # type: ignore[assignment]
+
+        return self._prepare_gen_out(text, finish, genconf)